[integralx][laravel] currency format extended to support editCategories. LedgerChartOfAccounts, BankTransactions, BankAccounts is ported to support currencyFormat. Fixes in pluggable format api

+++ laravel/resources/views/gridView.php
@@ -29,7 +29,7 @@
      grid model
      app as model
 
-     Last Modified: 16.03.2016
+     Last Modified: 17.03.2016
      Last Modified by: Nikita Zaharov
    -->
 
@@ -101,7 +101,12 @@
 						break;
 					    case "text":
 					    case "dropdown":
-						echo formatField($data->gridFields[$key], $value);
+						if(key_exists("formatFunction", $data->gridFields[$key])){
+						    $formatFunction = $data->gridFields[$key]["formatFunction"];
+						    echo $data->$formatFunction($row, "gridFields", $key, $value, false);
+						}
+						else
+						    echo formatField($data->gridFields[$key], $value);
 						break;
 					}
 					echo "</td>\n";
@@ -164,21 +169,27 @@
 			//renders table, contains record data using getEditItem from model
 			$item = $data->getEditItem($scope["item"], $scope["category"]);
 			foreach($item as $key =>$value){
-			    echo "<tr><td>" . $translation->translateLabel(key_exists($key, $data->columnNames) ? $data->columnNames[$key] : $key) . "</td><td>";
-			    switch($data->editCategories[$scope["category"]][$key]["inputType"]){
-				case "checkbox" :
-				    echo "<input class=\"grid-checkbox\" type=\"checkbox\"  ". ($value ? "checked" : "") . " disabled />";
-				    break;
-				case "timestamp" :
-				case "datetime" :
-				    echo date("m/d/y", strtotime($value));
-				    break;
-				case "text":
-				case "dropdown":
-				    echo formatField($data->editCategories[$scope["category"]][$key], $value);
-				    break;
+			    if(key_exists($key, $data->editCategories[$scope["category"]])){
+				echo "<tr><td>" . $translation->translateLabel(key_exists($key, $data->columnNames) ? $data->columnNames[$key] : $key) . "</td><td>";
+				switch($data->editCategories[$scope["category"]][$key]["inputType"]){
+				    case "checkbox" :
+					echo "<input class=\"grid-checkbox\" type=\"checkbox\"  ". ($value ? "checked" : "") . " disabled />";
+					break;
+				    case "timestamp" :
+				    case "datetime" :
+					echo date("m/d/y", strtotime($value));
+					break;
+				    case "text":
+				    case "dropdown":
+					if(key_exists("formatFunction", $data->editCategories[$scope["category"]][$key])){
+					    $formatFunction = $data->editCategories[$scope["category"]][$key]["formatFunction"];
+					    echo $data->$formatFunction($item, "editCategories", $key, $value, false);
+					}
+					else
+					    echo formatField($data->editCategories[$scope["category"]][$key], $value);						    break;
+				}
+				echo "</td></tr>";
 			    }
-			    echo "</td></tr>";
 			}
 			?>
 		    </tbody>
@@ -222,44 +233,53 @@
 		
 		foreach($item as $key =>$value){
 		    $translatedFieldName = $translation->translateLabel(key_exists($key, $data->columnNames) ? $data->columnNames[$key] : $key);
-		    switch($data->editCategories[$scope["category"]][$key]["inputType"]){
-			case "text" :
-			    //renders text input with label
-			    echo "<div class=\"form-group\"><label class=\"col-md-6\" for=\"" . $key ."\">" . $translatedFieldName . "</span></label><div class=\"col-md-6\"><input type=\"text\" id=\"". $key ."\" name=\"" .  $key. "\" class=\"form-control\" value=\"" . formatField($data->editCategories[$scope["category"]][$key], $value) ."\" " .
-				 ( (key_exists("disabledEdit", $data->editCategories[$scope["category"]][$key]) && $scope["mode"] == "edit")  || (key_exists("disabledNew", $data->editCategories[$scope["category"]][$key]) && $scope["mode"] == "new") ? "readonly" : "")
-				."></div></div>";
-			    break;
-			    
-			case "datetime" :
-			    //renders text input with label
-			    echo "<div class=\"form-group\"><label class=\"col-md-6\" for=\"" . $key ."\">" . $translatedFieldName . "</span></label><div class=\"col-md-6\"><input type=\"text\" id=\"". $key ."\" name=\"" .  $key. "\" class=\"form-control fdatetime\" value=\"" . ($value == 'now'? date("m/d/y") : date("m/d/y", strtotime($value))) ."\" " .
-				 ( (key_exists("disabledEdit", $data->editCategories[$scope["category"]][$key]) && $scope["mode"] == "edit")  || (key_exists("disabledNew", $data->editCategories[$scope["category"]][$key]) && $scope["mode"] == "new") ? "readonly" : "")
-				."></div></div>";
-			    break;
+		    if(key_exists($key, $data->editCategories[$scope["category"]])){
+			switch($data->editCategories[$scope["category"]][$key]["inputType"]){
+			    case "text" :
+				//renders text input with label
+				echo "<div class=\"form-group\"><label class=\"col-md-6\" for=\"" . $key ."\">" . $translatedFieldName . "</span></label><div class=\"col-md-6\"><input type=\"text\" id=\"". $key ."\" name=\"" .  $key. "\" class=\"form-control\" value=\"";
+				if(key_exists("formatFunction", $data->editCategories[$scope["category"]][$key])){
+				    $formatFunction = $data->editCategories[$scope["category"]][$key]["formatFunction"];
+				    echo $data->$formatFunction($item, "editCategories", $key, $value, false);
+				}
+				else
+				    echo formatField($data->editCategories[$scope["category"]][$key], $value);
 
-			case "checkbox" :
-			    //renders checkbox input with label
-			    echo "<input type=\"hidden\" name=\"" . $key . "\" value=\"0\"/>";
-			    echo "<div class=\"form-group\"><label class=\"col-md-6\" for=\"" . $key ."\">" . $translatedFieldName . "</span></label><div class=\"col-md-6\"><input class=\"grid-checkbox\" type=\"checkbox\" id=\"". $key ."\" name=\"" .  $key. "\" class=\"form-control\" value=\"1\" " . ($value ? "checked" : "") ." " .
-				 ( (key_exists("disabledEdit", $data->editCategories[$scope["category"]][$key]) && $scope["mode"] == "edit") || (key_exists("disabledNew", $data->editCategories[$scope["category"]][$key]) && $scope["mode"] == "new") ? "disabled" : "")
-				."></div></div>";
-			    break;
-			    
-			case "dropdown" :
-			    //renders select with available values as dropdowns with label
-			    echo "<div class=\"form-group\"><label class=\"col-sm-6\">" . $translatedFieldName . "</label><div class=\"col-sm-6\"><select class=\"form-control\" name=\"" . $key . "\" id=\"" . $key . "\">";
-			    $method = $data->editCategories[$scope["category"]][$key]["dataProvider"];
-			    $types = $data->$method();
-			    if($value)
-				echo "<option value=\"" . $value . "\">" . (key_exists($value, $types) ? $types[$value]["title"] : $value) . "</option>";
-			    else
-				echo "<option></option>";
+				echo"\" " . ( (key_exists("disabledEdit", $data->editCategories[$scope["category"]][$key]) && $scope["mode"] == "edit")  || (key_exists("disabledNew", $data->editCategories[$scope["category"]][$key]) && $scope["mode"] == "new") ? "readonly" : "")
+				    ."></div></div>";
+				break;
+				
+			    case "datetime" :
+				//renders text input with label
+				echo "<div class=\"form-group\"><label class=\"col-md-6\" for=\"" . $key ."\">" . $translatedFieldName . "</span></label><div class=\"col-md-6\"><input type=\"text\" id=\"". $key ."\" name=\"" .  $key. "\" class=\"form-control fdatetime\" value=\"" . ($value == 'now'? date("m/d/y") : date("m/d/y", strtotime($value))) ."\" " .
+				     ( (key_exists("disabledEdit", $data->editCategories[$scope["category"]][$key]) && $scope["mode"] == "edit")  || (key_exists("disabledNew", $data->editCategories[$scope["category"]][$key]) && $scope["mode"] == "new") ? "readonly" : "")
+				    ."></div></div>";
+				break;
 
-			    foreach($types as $type)
-				if(!$value || $type["value"] != $value)
-				    echo "<option value=\"" . $type["value"] . "\">" . $type["title"] . "</option>";
-			    echo"</select></div></div>";
-			    break;
+			    case "checkbox" :
+				//renders checkbox input with label
+				echo "<input type=\"hidden\" name=\"" . $key . "\" value=\"0\"/>";
+				echo "<div class=\"form-group\"><label class=\"col-md-6\" for=\"" . $key ."\">" . $translatedFieldName . "</span></label><div class=\"col-md-6\"><input class=\"grid-checkbox\" type=\"checkbox\" id=\"". $key ."\" name=\"" .  $key. "\" class=\"form-control\" value=\"1\" " . ($value ? "checked" : "") ." " .
+				     ( (key_exists("disabledEdit", $data->editCategories[$scope["category"]][$key]) && $scope["mode"] == "edit") || (key_exists("disabledNew", $data->editCategories[$scope["category"]][$key]) && $scope["mode"] == "new") ? "disabled" : "")
+				    ."></div></div>";
+				break;
+				
+			    case "dropdown" :
+				//renders select with available values as dropdowns with label
+				echo "<div class=\"form-group\"><label class=\"col-sm-6\">" . $translatedFieldName . "</label><div class=\"col-sm-6\"><select class=\"form-control\" name=\"" . $key . "\" id=\"" . $key . "\">";
+				$method = $data->editCategories[$scope["category"]][$key]["dataProvider"];
+				$types = $data->$method();
+				if($value)
+				    echo "<option value=\"" . $value . "\">" . (key_exists($value, $types) ? $types[$value]["title"] : $value) . "</option>";
+				else
+				    echo "<option></option>";
+
+				foreach($types as $type)
+				    if(!$value || $type["value"] != $value)
+					echo "<option value=\"" . $type["value"] . "\">" . $type["title"] . "</option>";
+				echo"</select></div></div>";
+				break;
+			}
 		    }
 		}
 		?>
